% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitEva.R
\name{FitEva}
\alias{FitEva}
\title{Evaluate the fitting result.}
\usage{
FitEva(Beta.simulate, Beta.fit, Gm, cutoff.sc, cutoff.g)
}
\arguments{
\item{Beta.simulate}{The simulated or golden standarded \code{B} matrix.}

\item{Beta.fit}{The fitted \code{B} matrix using \code{\link{PgaMsgl}} or other algorithms.}

\item{Gm}{Matrix of the group structure of coefficient matrix \code{B}. It is the a matrix of group boundaries, with each row indicating a group, four columns indicate the row-start, row-end, column-start and column-end of the group. The row/column index is 1-based.}

\item{cutoff.sc}{A number indicates the cutoff of a single coefficient in \code{B} to be true positive.}

\item{cutoff.g}{A number indicates the cutoff of the norm of a group (a sub-matrix in truth) in \code{B} to be true positive.}
}
\value{
\item{sc.TPR}{The true positive rate of single coefficients prediction.}

\item{sc.FPR}{The false positive rate of single coefficients prediction.}

\item{sc.TTPR}{The "true" true positive rate of single coefficients prediction.}

\item{sc.precision}{The precision of single coefficients prediction.}

\item{sc.accuracy}{The accuracy of single coefficients prediction.}

\item{sc.F1}{The F1 score of single coefficients prediction.}

\item{g.TPR}{The true positive rate of group prediction.}

\item{g.FPR}{The false positive rate of group prediction}

\item{g.precision}{The precision of goup prediction.}

\item{g.accuracy}{The accuracy of group prediction.}

\item{g.F1}{The F1 score of group prediction.}
}
\description{
Evaluate the fitting result by calculating the \strong{\code{TPR}} (ratio of true positives to real positives), \strong{\code{FPR}} (ratio of false positives to real negatives), \strong{\code{precision}} (ratio of true positives to predicted positives) , \strong{\code{accuracy}} (ratio of the sum of true positives and true negatives to total number of predictions) and \strong{\code{F1 score}} (the harmonic average of the precision and recall (TPR)).
}
\examples{

data(lowD)
result <- PgaMsgl(lowD$X, lowD$Y, lowD$B0, model="L121", lowD$Gm, lowD$mi, lowD$mg, lowD$mc)
FitEva_result <- FitEva(lowD$Beta, result$Beta, lowD$Gm, cutoff.sc=0, cutoff.g=0)

}
