% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highD.R
\docType{data}
\name{highD}
\alias{highD}
\title{Simulated example data on relative high dimension}
\format{A list with 18 components:
\describe{
\item{\strong{\code{X}}}{Simulated covariate matrix \code{X} in the model \code{Y=XB} with dimension 25x50, \emph{i.e.} \code{X} is composed with values of 50 variates in 25 samples. Elements in \code{X} are independently standard normal distributed.}

\item{\strong{\code{Beta0}}}{Simulated coefficient matrix \code{B} in the model \code{Y=XB} with dimension 50x2500. \code{B} is with 10x100 groups tiling it, and each group is a 5×25 module of \code{B}. 100 out of 1000 groups were randomly selected, each with 15 elements randomly selected to be with a random value in [-1, 1].}

\item{\strong{\code{Y}}}{Simulated response matrix \code{Y} in the model \code{Y=XB} with dimension 25x2500, i.e. \code{Y} is composed of values of 2500 responses in 25 samples. It is generated by multiply \code{X} with \code{Beta0}.}

\item{\strong{\code{Gm}}}{Matrix of the group structure of coefficient matrix \code{Beta0}. It is a 1000x4 matrix with each row indicating a group, four columns indicate the row-start, row-end, column-start and column-end of the group. The row/column index is 1-based.}

\item{\strong{\code{mi}}}{Maximum number of iteration, which is 2000.}

\item{\strong{\code{mg}}}{Maximum number of groups in matrix \code{B} to be reserved, which is 100.}

\item{\strong{\code{mc}}}{Maximum number of single coefficients in matrix \code{B} to be reserved, which is 1500.}

\item{\strong{\code{B0}}}{Simulated initial matrix of \code{B}, which is composed of all zeros.}

\item{\strong{\code{B1}}}{Simulated initial matrix of \code{B}, which is composed of all ones.}
 
\item{\strong{\code{B2}}}{Simulated initial matrix of \code{B}, which is generated adding random noise (rnorm(1,mean=1,sd=1)) to the real solution, Beta0.}

\item{\strong{\code{B3}}}{Simulated initial matrix of \code{B}, which is composed of all random noises (rnorm(1,mean=1,sd=1)).}

\item{\strong{\code{B4}}}{Simulated initial matrix of \code{B}, which is to have all non-zero elements in Beta0 as 1.}

\item{\strong{\code{B5}}}{Simulated initial matrix of \code{B}, which is to have non-zero values in Beta0 has 80\% probability to be 1, and zero values has 20\% probability to be 1.}

\item{\strong{\code{B6}}}{Simulated initial matrix of \code{B}, which is to have non-zero values in Beta0 has 80\% probability to be 1, and zero values to be 0.}

\item{\strong{\code{B7_001}}}{Simulated initial matrix of \code{B}. An element was set as 1 if the Benjamini-Hochberg adjusted P-value of the simple linear regression between the corresponding dependent variable and independent variable was less than 0.01, otherwise it was set as 0.}

\item{\strong{\code{B7_001_coeff}}}{Simulated initial matrix of \code{B}. An element was set as the coefficient of the simple linear regression between the corresponding dependent variable and independent variable, if its Benjamini-Hochberg adjusted P-value was less than 0.01, otherwise it was set as 0.}

\item{\strong{\code{B7_01}}}{Simulated initial matrix of \code{B}. Similar to \code{B7_001}, but the Benjamini-Hochberg adjusted P-value cutoff was 0.1.}

\item{\strong{\code{B7_01_coeff}}}{Simulated initial matrix of \code{B}. Similar to \code{B7_001_coeff}, but the Benjamini-Hochberg adjusted P-value cutoff was 0.1.}
}}
\description{
This is a list includes simulated example data on relative high dimension for \code{\link{PgaMsgl}} testing.
}
\examples{
data(highD)
system.time(highD_result <- PgaMsgl(highD$X, highD$Y, highD$B0, model="L121", highD$Gm, highD$mi, highD$mg, highD$mc))


}
\author{
Yiming Qin
}
\keyword{data}
