% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialBeta.R
\name{initialBeta}
\alias{initialBeta}
\title{Generate an initial Beta matrix for iteration with PgaMsgl.}
\usage{
initialBeta(XX, YY, method=p.adjust.methods, cutoff=p.adjust.cutoff, ncores=ncores)
}
\arguments{
\item{XX}{Matrix \code{X} in the model \code{Y=XB}.}

\item{YY}{Matrix \code{Y} in the model \code{Y=XB}.}

\item{method}{Method for p-value adjustment, acceptable values c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). More details see \code{\link{p.adjust}}.}

\item{cutoff}{A number between 0 and 1 indicates the cutoff of adjusted p-value.}

\item{ncores}{The number of cores to use for parallel execution. A parameter of \code{registerDoMC()} in \code{doMC} package.}
}
\value{
\item{B0}{The initial Beta matrix generated with simple linear regression with binary elements. The element in it is 1 if the simple linear regression of the corresponding dependent and independent variable passed the \code{p.adjust.cutoff}, and 0 if not.}

\item{B0_coeff}{The initial Beta matrix generated with simple linear regression. The element in it is the coefficient of the linear regression between variables if it passed the \code{p.adjust.cutoff}, and 0 if not.}
}
\description{
Generate an initial Beta matrix with simple linear regression of each dependent and independent variable.
}
\examples{

data(lowD)
B0 <- initialBeta(lowD$X, lowD$Y, "BH", 0.1, ncores=4)$B0
result <- PgaMsgl(lowD$X, lowD$Y, B0, model="L121", lowD$Gm, lowD$mi, lowD$mg, lowD$mc)
FitEva_result <- FitEva(lowD$Beta, result$Beta, lowD$Gm, cutoff.sc=0, cutoff.g=0)

}
