% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PgaMsgl.cv.R
\name{PgaMsgl.cv}
\alias{PgaMsgl.cv}
\title{Cross validation of super-parameters of PgaMsgl}
\usage{
PgaMsgl.cv(XX, YY, B0, model = c("L020v1", "L020v2", "L121"), Gm, mi = 1000, mg.v, mc.v, minlambda = 1e-5, rlambda = 0.98, mintau = 1e-5, rtau = 0.98, fold = 5, seed = 1, ncores)
}
\arguments{
\item{XX}{Matrix \code{X} in the model \code{Y=XB}. Remember to use \code{as.matrix()} if your input is a data frame.}

\item{YY}{Matrix \code{Y} in the model \code{Y=XB}. Remember to use \code{as.matrix()} if your input is a data frame.}

\item{B0}{Initial matrix of the coefficient matrix \code{B} in the model \code{Y=XB}. Remember to use \code{as.matrix()} if your input is a data frame.}

\item{model}{The model for Sparse Group Lasso, \code{L121}, \code{L020v1}, or \code{L020v2}.}

\item{Gm}{Matrix of the group structure of coefficient matrix \code{B}. It is the a matrix of group boundaries, with each row indicating a group, four columns indicate the row-start, row-end, column-start and column-end of the group. The row/column index is 1-based.}

\item{mi}{Maximum number of iterations allowed, default value is 1000.}

\item{mg.v}{A vector indicates maximum number of groups in matrix \code{B} to be reserved to be cross validated.}

\item{mc.v}{A vector indicates maximum number of single coefficients in matrix \code{B} to be reserved to be cross validated.}

\item{minlambda}{Minimum value of lambda. Only used when \code{model = "L020v2"}, default value is 1e-5.}

\item{rlambda}{Rate of lambda decrease. Only used when \code{model = "L020v2"}, defalult value is 0.98.}

\item{mintau}{Minimum value of tau. Only used when \code{model = "L020v2"}, default value is 1e-5.}

\item{rtau}{Rate of tau decrease. Only used when \code{model = "L020v2"}, default value is 0.98.}

\item{fold}{Number of fold for k-fold cross validation.}

\item{seed}{Numeric value for \code{set.seed()} when generate test or evaluate samples for k-fold cross validation.}

\item{ncores}{The number of cores to use for parallel execution. A parameter of \code{registerDoMC()} in \code{doMC} package.}
}
\value{
\item{rss}{A vector of the \code{residual sum of squares (RSS)} of model fitting with each combination of parameters.}

\item{RMSE}{A vector of the \code{root-mean-square error (RMSE)} of model fitting with each combination of parameters.}

\item{Rsquare}{A vector of the \eqn{R^{2}}{R^2} of model fitting with each combination of parameters.}

\item{rss.matr}{A matrix of the \code{residual sum of squares (RSS)} of model fitting with each combination of parameters.}

\item{RMSE.matr}{A matrix of the \code{root-mean-square error (RMSE)}E of model fitting with each combination of parameters.}

\item{Rsquare.matr}{A matrix of the \eqn{R^{2}}{R^2} of model fitting with each combination of parameters.}

\item{mg.v}{A vector of values of the parameter \code{mg} tested.}

\item{mc.v}{A vector of values of the parameter \code{mc} tested.}
}
\description{
Cross validation of super-parameters \code{mg} and \code{mc} of \code{\link{PgaMsgl}}.
}
\examples{
data(lowD)

mg.v <- seq(from=0.01*100, to=0.8*100, by=0.02*100)
mc.v <- seq(from = 0.01*2500, to = 0.5*2500, by = 0.01*2500)

result.cv <- PgaMsgl.cv(lowD$X, lowD$Y, lowD$B0, model="L121", lowD$Gm, lowD$mi, mg.v, mc.v, fold=5, seed=1, ncores=4)

grp.max.cv.rss <- result.cv$mg.v[which.min(result.cv$rss)]
coe.max.cv.rss <- result.cv$mc.v[which.min(result.cv$rss)]

grp.max.cv.rmse <- result.cv$mg.v[which.min(result.cv$RMSE)]
coe.max.cv.rmse <- result.cv$mc.v[which.min(result.cv$RMSE)]

grp.max.cv.r2 <- result.cv$mg.v[which.max(result.cv$Rsquare)]
coe.max.cv.r2 <- result.cv$mc.v[which.max(result.cv$Rsquare)]

system.time(try1 <- PgaMsgl(lowD$X, lowD$Y, lowD$B0, model="L121", lowD$Gm, lowD$mi, 10, 120))
system.time(try2 <- PgaMsgl(lowD$X, lowD$Y, lowD$B0, model="L121", lowD$Gm, lowD$mi, grp.max.cv.rss, coe.max.cv.rss))
system.time(try3 <- PgaMsgl(lowD$X, lowD$Y, lowD$B0, model="L121", lowD$Gm, lowD$mi, grp.max.cv.rmse, coe.max.cv.rmse))
system.time(try4 <- PgaMsgl(lowD$X, lowD$Y, lowD$B0, model="L121", lowD$Gm, lowD$mi, grp.max.cv.r2, coe.max.cv.r2))

}
\author{
Yiming Qin
}
