% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PgaMsgl.R
\name{PgaMsgl}
\alias{PgaMsgl}
\title{Proximal Gradient Algorithm for Multi-variate Sparse Group Lasso}
\usage{
PgaMsgl(XX, YY, B0, model = c("L020v1", "L020v2", "L121"), Gm, mi = 1000, mg, mc, minlambda = 1e-5, rlambda = 0.98, mintau = 1e-5, rtau = 0.98)
}
\arguments{
\item{XX}{Matrix \code{X} in the model \code{Y=XB}.}

\item{YY}{Matrix \code{Y} in the model \code{Y=XB}.}

\item{B0}{Initial matrix of the coefficient matrix \code{B} in the model \code{Y=XB}.}

\item{model}{The model for Sparse Group Lasso, \code{L121}, \code{L020v1}, or \code{L020v2}.}

\item{Gm}{Matrix of the group structure of coefficient matrix \code{B}. It is the a matrix of group boundaries, with each row indicating a group, four columns indicate the row-start, row-end, column-start and column-end of the group. The row/column index is 1-based.}

\item{mi}{Maximum number of iterations allowed, default value is 1000.}

\item{mg}{An interger indicates maximum number of groups in matrix \code{B} to be reserved.}

\item{mc}{An interger indicates maximum number of single coefficients in matrix \code{B} to be reserved.}

\item{minlambda}{Minimum value of lambda. Only used when \code{model = "L020v2"}, default value is 1e-5.}

\item{rlambda}{Rate of lambda decrease. Only used when \code{model = "L020v2"}, defalult value is 0.98.}

\item{mintau}{Minimum value of tau. Only used when \code{model = "L020v2"}, default value is 1e-5.}

\item{rtau}{Rate of tau decrease. Only used when \code{model = "L020v2"}, default value is 0.98.}
}
\value{
\item{Beta}{The estimated coefficient matrix \code{B} in the model \code{Y=XB}.}

\item{Rss}{A vector with length \code{mi} (maximum number of iterations), which is the \code{residual sum of squares (RSS)} at each iteration.}

\item{Tau}{A vector with length \code{mi} (maximum number of iterations), which is the value of parameter \code{tau} at each iteration step.}

\item{Lambda}{A vector with length \code{mi} (maximum number of iterations), which is the value of parameter \code{lambda} at each iteration step.}

\item{iteration.time}{An interger, which is the times of iterations in practice.}

\item{Rss_relative}{A vector with length \code{mi} (maximum number of iterations), which is \code{Rss} divided by square of \code{Y}'s norm.}
}
\description{
Proximal Gradient Algorithm for Multi-variate Sparse Group Lasso.
}
\examples{
data(lowD)
result <- PgaMsgl(lowD$X, lowD$Y, lowD$B0, model="L121", lowD$Gm, lowD$mi, lowD$mg, lowD$mc)

}
