% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimMax.R
\name{SimMax}
\alias{SimMax}
\title{Simulate matrices X, B, and Y in the model Y=XB.}
\usage{
SimMax(N, p, q, n.g.row, n.g.col, mg, mc, seed)
}
\arguments{
\item{N}{Number of samples to be simulated, which is also the number of rows in matrices \code{X} and \code{Y}.}

\item{p}{Number of covariates, which is also the number of columns of \code{X} or number of rows of \code{B}.}

\item{q}{Number of response variables, which is also the number of columns of \code{Y} and \code{B}.}

\item{n.g.row}{Number of groups of rows in \code{B}. It should be divisible by \code{p}.}

\item{n.g.col}{Number of groups of columns in \code{B}. It should be divisible by \code{q}.}

\item{mg}{An interger indicates maximum number of non-zero groups in matrix \code{B}.}

\item{mc}{An interger indicates maximum number of non-zero single coefficients in matrix \code{B}.}

\item{seed}{Numeric value for \code{set.seed()} for reproduce the random latter.}
}
\value{
\item{X}{Simulated matrix \code{X} with dimension \code{Nxp} in the model \code{Y=XB}.}

\item{Beta}{Simulate coefficient matrix \code{B} with dimension \code{pxq} in the model \code{Y=XB}.}

\item{Y}{Simulated matrix \code{Y} with dimension \code{Nxq} in the model \code{Y=XB}.}

\item{Gm}{Matrix of the group structure of coefficient matrix \code{B}. It is the a matrix of group boundaries, with each row indicating a group, four columns indicate the row-start, row-end, column-start and column-end of the group. The row/column index is 1-based.}
}
\description{
Simulate matrices \code{X}, \code{B}, and \code{Y} in the model \code{Y=XB} with a certain group structure and sparsity of \code{B}. With this function, groups can only be simulated with equal sizes.
}
\examples{

N <- 25
p <- 50
q <- 50
n.g.row <- 5
n.g.col <- 5
mg <- 10
mc <- 100
seed <- 1
sim <- SimMax(N, p, q, n.g.row, n.g.col, mg, mc, seed)

}
